Remove unnecessary dependencies EC2/SQS
